+++ lang/syn/src/lib.rs
@@ -613,6 +613,7 @@ pub struct ConstraintSigner {}
 #[derive(Debug, Clone)]
 pub struct ConstraintHasOne {
     pub join_target: Expr,
+    pub error: Option<Expr>,
 }
 
 #[derive(Debug, Clone)]